+++ src/core/component/directives/safe-html/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.? (2024-??-??)
+
+#### :rocket: New Feature
+
+* Supported the insertion of a primitive as a value
+
 ## v3.68.0 (2024-03-04)
 
 #### :rocket: New Feature